test: merge AssertContains asserts
It slightly modifies the behavior of AssertContains as it now checks for
all properties of ExpectedDifference at once instead of splitting the
rule and jsonref checks into two separate asserts. This is more in line
with what we expect from the method.

AssertContainsOnly is replaced by an optional param to AssertContains.